requirements.txt
Django==5.2.4
Pillow>=9.0.0
django-ckeditor>=6.7.0
django-taggit>=5.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
openai>=1.0.0
groq>=0.4.1
numpy>=1.24.0
tiktoken>=0.5.1
psycopg2-binary>=2.9.0
dj-database-url
openai>=1.0.0
